# (dev) Pre-generate crypto materials