--- a/charts/hlf-k8s/templates/deployment-enrollement-operator.yaml
+++ b/charts/hlf-k8s/templates/deployment-enrollement-operator.yaml
@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+{{- if index .Values "hlf-ca" "enabled" }}
 apiVersion: apps/v1
 kind: Deployment
 metadata:
@@ -167,3 +168,4 @@ spec:
       nodeSelector:
         {{- toYaml . | nindent 8 }}
       {{- end }}
+{{- end}}