--- a/charts/hlf-k8s/templates/deployment-genesis-operator.yaml
+++ b/charts/hlf-k8s/templates/deployment-genesis-operator.yaml
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 {{- if index .Values "hlf-ord" "enabled" }}
+{{- if .Values.genesis.generate }}
 apiVersion: apps/v1
 kind: Deployment
 metadata:
@@ -142,4 +143,5 @@ spec:
       tolerations:
         {{- toYaml . | nindent 8 }}
     {{- end }}
+{{- end }}
 {{- end }}