--- a/examples/2-orgs-policy-any-no-ca/skaffold.yaml
+++ b/examples/2-orgs-policy-any-no-ca/skaffold.yaml
@@ -17,7 +17,7 @@ kind: Config
 build:
   artifacts:
     - image: substrafoundation/hlf-k8s
-      context: .
+      context: ../../
       docker:
         dockerfile: images/hlf-k8s/Dockerfile
       sync:
@@ -30,18 +30,18 @@ deploy:
   helm:
     releases:
       - name: network-orderer
-        chartPath: charts/hlf-k8s
+        chartPath: ../../charts/hlf-k8s
         namespace: orderer
-        valuesFiles: [examples/2-orgs-policy-any-no-ca/values/orderer.yaml]
+        valuesFiles: [values/orderer.yaml]
         imageStrategy:
           helm: {}
         values:
           image: substrafoundation/hlf-k8s
 
       - name: network-org-1-peer-1
-        chartPath: charts/hlf-k8s
+        chartPath: ../../charts/hlf-k8s
         namespace: org-1
-        valuesFiles: [examples/2-orgs-policy-any-no-ca/values/org-1-peer-1.yaml]
+        valuesFiles: [values/org-1-peer-1.yaml]
         imageStrategy:
           helm: {}
         values:
@@ -49,9 +49,9 @@ deploy:
 
 
       - name: network-org-2-peer-1
-        chartPath: charts/hlf-k8s
+        chartPath: ../../charts/hlf-k8s
         namespace: org-2
-        valuesFiles: [examples/2-orgs-policy-any-no-ca/values/org-2-peer-1.yaml]
+        valuesFiles: [values/org-2-peer-1.yaml]
         imageStrategy:
           helm: {}
         values:
@@ -67,3 +67,4 @@ deploy:
       - examples/secrets/secrets-orderer.yaml
       - examples/secrets/secrets-org-1.yaml
       - examples/secrets/secrets-org-2.yaml
+