--- a/examples/2-orgs-policy-any-no-ca/values/orderer.yaml
+++ b/examples/2-orgs-policy-any-no-ca/values/orderer.yaml
@@ -1,0 +1,51 @@
+# Orderer values for 1 orderer 2 organizations setup
+
+organization:
+  id: MyOrdererMSP
+  name: MyOrderer
+
+nginx-ingress:
+  enabled: true
+
+hlf-peer:
+  enabled: false
+
+hlf-ca:
+  enabled: false
+  caName: rcaOrderer
+  host: network-orderer-hlf-ca.orderer
+  orderer:
+    host: network-orderer-hlf-ca.orderer
+
+hlf-ord:
+  enabled: true
+  monitor:
+    enabled: true
+  host: network-orderer-hlf-ord.orderer
+  ord:
+    mspID: MyOrdererMSP
+
+appChannels:
+  - channelName: mychannel
+  - channelName: yourchannel
+
+systemChannel:
+  organizations:
+    - { org: MyOrg1, mspid: MyOrg1MSP, configUrl: network-org-1-peer-1-hlf-k8s-config-operator.org-1/config/configOrg.json }
+    - { org: MyOrg2, mspid: MyOrg2MSP, configUrl: network-org-2-peer-1-hlf-k8s-config-operator.org-2/config/configOrg.json }
+
+enrollments:
+  creds:
+  - { name: admin, secret: adminpwd, options: "--id.attrs admin=true:ecert"}
+  - { name: user, secret: pwd, options: "--id.type orderer"}
+  csrHost: network-orderer-hlf-ord.orderer
+
+toolbox:
+  enabled: true
+
+hooks:
+  deleteSecrets:
+    enabled: false
+
+genesis:
+  generate: false