--- a/examples/2-orgs-policy-any-no-ca/values/org-2-peer-1.yaml
+++ b/examples/2-orgs-policy-any-no-ca/values/org-2-peer-1.yaml
@@ -1,0 +1,61 @@
+organization:
+  id: MyOrg2MSP
+  name: MyOrg2
+
+nginx-ingress:
+  enabled: true
+  controller:
+    scope:
+      enabled: true
+
+
+hlf-ca:
+  enabled: false
+  caName: rcaOrg2
+  host: network-org-2-peer-1-hlf-ca.org-2
+  orderer:
+    host: network-orderer-hlf-ca.orderer
+
+hlf-ord:
+  host: network-orderer-hlf-ord.orderer
+
+hlf-peer:
+  host: network-org-2-peer-1-hlf-peer.org-2
+  peer:
+    mspID: MyOrg2MSP
+    gossip:
+      externalEndpoint: network-org-2-peer-1-hlf-peer.org-2:7051
+  discover-monitor:
+    enabled: true
+
+
+chaincodes:
+  - name: mycc
+    version: "1.0"
+    src: https://github.com/SubstraFoundation/substra-chaincode/archive/master.tar.gz
+
+appChannels:
+  - channelName: mychannel
+    chaincodePolicy: "OR('MyOrg1MSP.member','MyOrg2MSP.member')"
+    chaincodeName: mycc
+    chaincodeVersion: "1.0"
+
+  - channelName: yourchannel
+    chaincodePolicy: "OR('MyOrg1MSP.member','MyOrg2MSP.member')"
+    chaincodeName: mycc
+    chaincodeVersion: "1.0"
+
+
+enrollments:
+  creds:
+    - { name: admin, secret: adminpwd, options: "--id.attrs hf.Registrar.Roles=client,hf.Registrar.Attributes=*,hf.Revoker=true,hf.GenCRL=true,admin=true:ecert,abac.init=true:ecert"}
+    - { name: user, secret: pwd, options: "--id.type peer"}
+  csrHost: network-org-2-peer-1-hlf-peer.org-2
+
+
+toolbox:
+  enabled: true
+
+hooks:
+  deleteSecrets:
+    enabled: false