… examples/secrets/secrets-orderer-genesis.yaml
… examples/secrets/secrets-orderer.yaml
… examples/secrets/secrets-org-1.yaml
… examples/secrets/secrets-org-2.yaml
… examples/secrets/secrets-org-3.yaml
… examples/secrets/secrets-org-4.yaml