Companion PR: https://github.com/SubstraFoundation/substra-tests/pull/151

# Overview 

Pre-generate HLF crypto materials to speedup local development:

- no enrollment pod
- no genesis pod
- orderer / peer pods start instantly

# Benchmarks (local env)

|  | **Total time to readiness** | skaffold run | Deployment stabilized | Chaincode instantiated | skaffold delete | 
|--|--|--|--|--|--|
| Without pre-gen | **143s**  | 100s | 39s | 43s | 25s |
| With pre-gen | **80s** | 60s | 5s | 20s | 13s | 

Note: on the CI, there's a big speedup in deployment stabilization time (from ~6 min to ~2 min)

# How to use

## Deploy

### Without pre-gen

```bash
$ skaffold run
```

### With pre-gen

```bash
$ ./examples/secrets-create.sh
$ skaffold run
```

## Delete

```
$ skaffold delete
```

Secrets will be deleted regardless of pre-gen status. To re-deploy with pre-gen, you need to run `examples/secrets-create.sh` again.

To delete secrets only, run `./examples/secrets-create.sh delete`

# Limitations and Future work

Skaffold deployment order is not configurable: by default, helm charts are deployed _before_ kubectl manifest.

So if the user doesn't  run `examples/secrets-create.sh` before `skaffold run`, skaffold will still create the secrets, but **after** deploying the helm charts. This greatly limits the speedup benefits. 

There is a [skaffold issue](https://github.com/GoogleContainerTools/skaffold/issues/3905) opened on the topic.  Once it's addressed, we can get rid of `examples/secrets-create.sh` altogether.