@Kelvin-M 2020-10-01
I found `./examples/secrets-create.sh delete` pretty strange, maybe it would be better to have something like : 

```
./examples/hlf-secrets.sh create
./examples/hlf-secrets.sh delete
```