@Kelvin-M 2020-10-01
Moreover, we should warn the user that's only for dev setup and should not be used in production :D !