@AurelienGasser 2020-10-01
@Kelvin-M What about

```
./examples/sample-secrets.sh create
./examples/sample-secrets.sh delete
```

Would that address both your concerns?