@AurelienGasser 2020-10-01
or `dev-secrets.sh` ?